+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 0.24 (2024-09-??)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/15)
-
+* Develop parent visibility for entities
 
 0.24 (2024-09-04)
 ===============